DeathStarBench/hotelReservation/services-python/reservation/requirements.txt
pymongo
python-memcached
requests